pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag-qa-system"
version = "0.1.0"
description = "Agentic RAG QA system with LangGraph + LangChain"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "langchain>=0.3.0",
  "langgraph>=0.2.0",
  "langchain-community>=0.3.0",
  "langchain-chroma>=0.1.0",
  "langchain-huggingface>=0.1.0",
  "langchain-deepseek>=0.1.0",
  "chromadb>=0.5.0",
  "sentence-transformers>=3.0.0",
  "pypdf>=4.2.0",
  "orjson>=3.10.0",
  "tqdm>=4.66.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
